added about tab that links to previous home page, made new home page PennApps XXV: Coming Soon...
